Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@
 int main() {
     int num1, num2, num3;
     scanf("%d %d %d", &num1, &num2, &num3);
-    int avg = (num1 + num2+ num3)/3.0;
-    printf("Average: %.2d", avg);
+    double avg = (num1 + num2 + num3)/3.0;
+    printf("Average: %.2lf", avg);
     return 0;
 }
